exercise-files/02/02_01/requirements.txt
colorama==0.4.4
python-dotenv==1.0.0
openai
requests==2.31.0

